+++ src/main/java/qikahome/jsonmore/lib/recipe/IConsumingRecipe.java
@@ -1,5 +1,6 @@
 package qikahome.jsonmore.lib.recipe;
 
+// TODO 2.0: Rename?
 public interface IConsumingRecipe {
     
 }